+++ api/src/modules/moderation/moderation.controller.e2e-test.ts
@@ -2,6 +2,7 @@ import { getRepository, Repository } from 'typeorm';
 
 import { createAuthenticatedModerator, createAuthenticatedUser, setupE2eTest } from '../../testing/setup-e2e-test';
 import { AuthenticationModule } from '../authentication/authentication.module';
+import { Comment } from '../comment/comment.entity';
 import { CommentFactory } from '../comment/comment.factory';
 import { CommentModule } from '../comment/comment.module';
 import { Report, ReportModerationAction } from '../comment/report/report.entity';
@@ -17,6 +18,7 @@ describe('moderation', () => {
     imports: [UserModule, AuthenticationModule, CommentModule, ModerationModule],
   });
 
+  let commentRepository: Repository<Comment>;
   let reportRepository: Repository<Report>;
 
   let createUser: UserFactory['create'];
@@ -37,6 +39,7 @@ describe('moderation', () => {
     createComment = commentFactory.create.bind(commentFactory);
     reportComment = reportFactory.create.bind(reportFactory);
 
+    commentRepository = getRepository(Comment);
     reportRepository = getRepository(Report);
   });
 
@@ -125,6 +128,46 @@ describe('moderation', () => {
         moderationAction: ReportModerationAction.IGNORED,
         moderatedBy: expect.objectContaining({ id: moderator.id }),
       });
+    });
+
+  });
+
+  describe('delete reported comment', () => {
+
+    it('should not delete a comment when not a moderator', async () => {
+      const comment = await createComment();
+      await reportComment({ reporter: await createUser(), comment });
+
+      await asUser.put('/api/moderation/delete-comment')
+        .send({ commentId: comment.id })
+        .expect(403);
+    });
+
+    it('should delete a comment', async () => {
+      const comment = await createComment();
+      const report1 = await reportComment({ reporter: await createUser(), comment });
+      const report2 = await reportComment({ reporter: await createUser(), comment });
+
+      await asModerator
+        .put('/api/moderation/delete-comment')
+        .send({ commentId: comment.id })
+        .expect(204);
+
+      const [reportDb1, reportDb2] = await reportRepository.findByIds([report1.id, report2.id]);
+
+      expect(reportDb1).toMatchObject({
+        moderationAction: ReportModerationAction.DELETED,
+        moderatedBy: expect.objectContaining({ id: moderator.id }),
+      });
+
+      expect(reportDb2).toMatchObject({
+        moderationAction: ReportModerationAction.DELETED,
+        moderatedBy: expect.objectContaining({ id: moderator.id }),
+      });
+
+      const commentDb = await commentRepository.findOne({ id: comment.id }, { withDeleted: true });
+
+      expect(commentDb).toMatchObject({ deleted: expect.any(Date) });
     });
 
   });