+++ api/src/modules/moderation/moderation.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, HttpCode, HttpStatus, Put, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Get, HttpCode, HttpStatus, NotFoundException, Put, UseGuards, UseInterceptors } from '@nestjs/common';
 
 import { AuthUser } from '../../common/auth-user.decorator';
 import { IsAuthenticated } from '../../common/auth.guard';
@@ -8,10 +8,12 @@ import { PageQuery } from '../../common/page-query.decorator';
 import { Paginated } from '../../common/paginated';
 import { Roles } from '../../common/roles.decorator';
 import { Role } from '../authorization/roles.enum';
+import { CommentService } from '../comment/comment.service';
 import { PopulateComment } from '../comment/populate-comment.interceptor';
 import { ReportedCommentDto } from '../comment/report/dtos/reported-comment-dto';
 import { User } from '../user/user.entity';
 
+import { DeleteCommentInDto } from './dtos/delete-comment-in.dto';
 import { IgnoreReportsInDto } from './dtos/ignore-reports-in.dto';
 import { CommentWithReports, ModerationService } from './moderation.service';
 
@@ -20,12 +22,13 @@ import { CommentWithReports, ModerationService } from './moderation.service';
 export class ModerationController {
 
   constructor(
+    private readonly commentService: CommentService,
     private readonly moderationService: ModerationService,
   ) {}
 
   @Get('reports')
   @UseGuards(IsAuthenticated)
-  @Roles(Role.MODERATOR)
+  @Roles(Role.MODERATOR, Role.ADMIN)
   @UseInterceptors(PopulateComment)
   @CastToDto(ReportedCommentDto)
   async findCommentsWaitingForReview(
@@ -38,13 +41,34 @@ export class ModerationController {
 
   @Put('ignore-reports')
   @UseGuards(IsAuthenticated)
-  @Roles(Role.MODERATOR)
+  @Roles(Role.MODERATOR, Role.ADMIN)
   @HttpCode(HttpStatus.NO_CONTENT)
   async ignoreReports(
     @Body() { commentId }: IgnoreReportsInDto,
     @AuthUser() user: User,
   ): Promise<void> {
-    await this.moderationService.ignoreReports(commentId, user);
+    const comment = await this.commentService.findById(commentId);
+
+    if (!comment)
+      throw new NotFoundException();
+
+    await this.moderationService.ignoreReports(comment, user);
+  }
+
+  @Put('delete-comment')
+  @UseGuards(IsAuthenticated)
+  @Roles(Role.MODERATOR, Role.ADMIN)
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async deleteComment(
+    @Body() { commentId }: DeleteCommentInDto,
+    @AuthUser() user: User,
+  ): Promise<void> {
+    const comment = await this.commentService.findById(commentId);
+
+    if (!comment)
+      throw new NotFoundException();
+
+    await this.moderationService.deleteComment(comment, user);
   }
 
 }