+++ app/src/theme/createTheme.ts
@@ -27,7 +27,7 @@ const createTheme = (): Theme => {
       highlight: { main: pink[500] },
       border: { main: '#ccc', light: '#ddd', veryLight: '#eee' },
       selected: { main: '#ffeeaa' },
-      textLight: { main: '#666666', light: '#999999' },
+      textLight: { main: 'rgba(0, 0, 0, 0.7)', light: 'rgba(0, 0, 0, 0.6)', dark: 'rgba(0, 0, 0, 0.8)' },
       textLink: { main: '#44C', light: '#55A', dark: '#446' },
       textWarning: { main: '#ee7700' },
       background: { default: 'white' },